cryptos/requirements.txt
krakenex
matplotlib
numpy
pandas
pyyaml
scikit-learn
scipy
streamlit>=1.1
tqdm
python-dotenv
python-kraken-sdk